services/ai-engine/requirements.txt
# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=24.1.0

# Database
supabase>=2.7.0

# AI/LLM
groq>=0.9.0
openai>=1.0.0

# WhatsApp + SMS
twilio>=9.0.0

# Search
tavily-python>=0.7.0

# Utils
numpy>=1.26.0
python-jose>=3.3.0
Pillow>=10.4.0
